feat: relocate component registration
